exclusion of testcases that require testcases and other modifications
