+++ README.md
@@ -66,7 +66,27 @@ The `$user_prefix` is usually your dockerhub user id.
 
 ### Testing
 
+Note: If you're running all tests locally (either of these two commands: `./gradlew tests:test` or `./gradlew tests:test --tests *CredentialsIBMNodeJsActionWatsonTests`), you need to set up a tests/vcap_services.json file containing Watson credentials in the format of:
+```  
+{  
+  "language_translation":[  
+    {  
+      "name":"",  
+      "label":"",  
+      "plan":"",  
+      "credentials": {  
+        "url": "",  
+        "password": "",  
+        "username": ""  
+      }  
+    }  
+  ]  
+}  
+```  
+
 To run all tests: `./gradlew tests:test`
+
+To run all tests except those which do not rely on credentials `./gradlew tests:testWithoutCredentials`
 
 To run a single test-class: `./gradlew tests:test --tests <SomeGradleTestFilter>`
